Split code in packaged structure
Work on #2
