+++ src/main/resources/application.yml
@@ -1,10 +1,26 @@
-spring:
-  ai:
-    gigachat:
-      scope: GIGACHAT_API_PERS
-      client-id: ${SPRING_AI_GIGACHAT_CLIENT_ID}
-      client-secret: ${SPRING_AI_GIGACHAT_CLIENT_SECRET}
-      unsafe-ssl: true
-      chat:
-        options:
-          model: GigaChat-2-Max
+spring.ai.gigachat:
+  scope: ${SPRING_AI_GIGACHAT_SCOPE:GIGACHAT_API_CORP}
+  client-id: ${SPRING_AI_GIGACHAT_CLIENT_ID}
+  client-secret: ${SPRING_AI_GIGACHAT_CLIENT_SECRET}
+  unsafe-ssl: ${SPRING_AI_GIGACHAT_UNSAFE_SSL:true}
+  chat:
+    options:
+      model: GigaChat-2-Max
+      temperature: 0.5
+      top-p: 0.5
+      max-tokens: 1000
+
+management:
+  tracing:
+    sampling:
+      probability: 1
+  otlp:
+    tracing:
+      export:
+        enabled: ${MANAGEMENT_OTLP_TRACING_EXPORT_ENABLED:false}
+      transport: ${MANAGEMENT_OTLP_TRACING_TRANSPORT:http}
+      endpoint: ${MANAGEMENT_OTLP_TRACING_ENDPOINT:localhost:8080}
+  endpoint:
+    health:
+      probes:
+        enabled: true
